Add a service CSRF protection and deprecate event listener methods - related to #7611
