+++ bundles/AdminBundle/EventListener/CsrfProtectionListener.php
@@ -14,40 +14,29 @@
 
 namespace Pimcore\Bundle\AdminBundle\EventListener;
 
+use Pimcore\Bundle\AdminBundle\Security\CsrfProtectionHandler;
 use Pimcore\Bundle\CoreBundle\EventListener\Traits\PimcoreContextAwareTrait;
 use Pimcore\Http\Request\Resolver\PimcoreContextResolver;
-use Pimcore\Templating\PhpEngine;
-use Pimcore\Tool\Session;
-use Psr\Log\LoggerAwareTrait;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Session\Attribute\AttributeBagInterface;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
-use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\HttpKernel\KernelEvents;
 
 class CsrfProtectionListener implements EventSubscriberInterface
 {
     use PimcoreContextAwareTrait;
-    use LoggerAwareTrait;
-
-    protected $excludedRoutes = [];
-
-    protected $csrfToken = null;
 
     /**
-     * @var PhpEngine
+     * @var CsrfProtectionHandler $handler
      */
-    protected $phpTemplatingEngine;
+    protected $csrfProtectionHandler;
 
     /**
-     * @param array $excludedRoutes
-     * @param PhpEngine $phpTemplatingEngine
+     * @param CsrfProtectionHandler $csrfProtectionHandler
      */
-    public function __construct($excludedRoutes, PhpEngine $phpTemplatingEngine)
+    public function __construct(CsrfProtectionHandler $csrfProtectionHandler)
     {
-        $this->excludedRoutes = $excludedRoutes;
-        $this->phpTemplatingEngine = $phpTemplatingEngine;
+        $this->csrfProtectionHandler = $csrfProtectionHandler;
     }
 
     /**
@@ -70,7 +59,7 @@ class CsrfProtectionListener implements EventSubscriberInterface
             return;
         }
 
-        $this->phpTemplatingEngine->addGlobal('csrfToken', $this->getCsrfToken());
+        $this->csrfProtectionHandler->generateCsrfToken();
 
         if ($request->getMethod() == Request::METHOD_GET) {
             return;
@@ -88,63 +77,49 @@ class CsrfProtectionListener implements EventSubscriberInterface
         ];
 
         $route = $request->attributes->get('_route');
-        if (in_array($route, $exludedRoutes) || in_array($route, $this->excludedRoutes)) {
+        if (in_array($route, $exludedRoutes) || in_array($route, $this->csrfProtectionHandler->getExcludedRoutes())) {
             return;
         }
 
-        $this->checkCsrfToken($request);
+        $this->csrfProtectionHandler->checkCsrfToken($request);
     }
 
     /**
      * @param Request $request
+     *
+     * @deprecated use CsrfProtectionHandler::checkCsrfToken() instead
      */
     public function checkCsrfToken(Request $request)
     {
-        $csrfToken = $this->getCsrfToken();
-        $requestCsrfToken = $request->headers->get('x_pimcore_csrf_token');
-        if (!$requestCsrfToken) {
-            $requestCsrfToken = $request->get('csrfToken');
-        }
+        @trigger_error(sprintf('Calling '.__METHOD__.' is deprecated since version 6.9.0 and will be removed in 7.0.0. ' .
+            'Use %s service instead.', CsrfProtectionHandler::class), E_USER_DEPRECATED);
 
-        if (!$csrfToken || $csrfToken !== $requestCsrfToken) {
-            $this->logger->error('Detected CSRF attack on {request}', [
-                'request' => $request->getPathInfo(),
-            ]);
-
-            throw new AccessDeniedHttpException('Detected CSRF Attack! Do not do evil things with pimcore ... ;-)');
-        }
+        $this->csrfProtectionHandler->checkCsrfToken($request);
     }
 
     /**
      * @return string
+     *
+     * @deprecated use CsrfProtectionHandler::getCsrfToken() instead
      */
     public function getCsrfToken()
     {
-        if (!$this->csrfToken) {
-            $this->csrfToken = Session::getReadOnly()->get('csrfToken');
-            if (!$this->csrfToken) {
-                $this->csrfToken = Session::useSession(function (AttributeBagInterface $adminSession) {
-                    if (!$adminSession->has('csrfToken') && !$adminSession->get('csrfToken')) {
-                        $adminSession->set('csrfToken', sha1(generateRandomSymfonySecret()));
-                    }
+        @trigger_error(sprintf('Calling '.__METHOD__.' is deprecated since version 6.9.0 and will be removed in 7.0.0. ' .
+            'Use %s service instead.', CsrfProtectionHandler::class), E_USER_DEPRECATED);
 
-                    return $adminSession->get('csrfToken');
-                });
-            }
-        }
-
-        return $this->csrfToken;
+        return $this->csrfProtectionHandler->getCsrfToken();
     }
 
+    /**
+     * @return string
+     *
+     * @deprecated use CsrfProtectionHandler::getCsrfToken() instead
+     */
     public function regenerateCsrfToken()
     {
-        $this->csrfToken = Session::useSession(function (AttributeBagInterface $adminSession) {
-            $token = sha1(generateRandomSymfonySecret());
-            $adminSession->set('csrfToken', $token);
+        @trigger_error(sprintf('Calling '.__METHOD__.' is deprecated since version 6.9.0 and will be removed in 7.0.0. ' .
+            'Use %s service instead.', CsrfProtectionHandler::class), E_USER_DEPRECATED);
 
-            return $token;
-        });
-
-        $this->phpTemplatingEngine->addGlobal('csrfToken', $this->csrfToken);
+        $this->csrfProtectionHandler->regenerateCsrfToken();
     }
 }
